apps/api/requirements.txt
# FastAPI and server
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
prisma==0.12.0

# Authentication
PyJWT==2.8.0
passlib[bcrypt]==1.7.4

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Email (for future Proton Mail integration)
aiosmtplib==3.0.1

# HTTP client
httpx==0.26.0

# Utilities
python-dotenv==1.0.1
